refactor: move Auth parsing to ilp-service
