+++ crates/interledger-btp/Cargo.toml
@@ -16,7 +16,6 @@ futures = "0.1.25"
 interledger-ildcp = { path = "../interledger-ildcp", version = "0.2.1" }
 interledger-packet = { path = "../interledger-packet", version = "0.2.1" }
 interledger-service = { path = "../interledger-service", version = "0.2.1" }
-interledger-http = { path = "../interledger-http", version = "0.2.1" }
 log = "0.4.6"
 num-bigint = "0.2.2"
 parking_lot = "0.7.1"
