+++ crates/interledger-http/src/lib.rs
@@ -2,8 +2,6 @@
 //!
 //! Client and server implementations of the [ILP-Over-HTTP](https://github.com/interledger/rfcs/blob/master/0035-ilp-over-http/0035-ilp-over-http.md) bilateral communication protocol.
 //! This protocol is intended primarily for server-to-server communication between peers on the Interledger network.
-
-use base64;
 use futures::Future;
 use interledger_service::Account;
 use url::Url;
@@ -34,105 +32,4 @@ pub trait HttpStore: Clone + Send + Sync + 'static {
         username: &str,
         token: &str,
     ) -> Box<dyn Future<Item = Self::Account, Error = ()> + Send>;
-}
-
-// Helpers for parsing authorization methods
-pub const BEARER_TOKEN_START: usize = 7;
-pub const BASIC_TOKEN_START: usize = 6;
-
-#[derive(Eq, PartialEq, Debug, Clone)]
-pub struct Auth {
-    username: String,
-    password: String,
-}
-
-impl Auth {
-    pub fn new(username: &str, password: &str) -> Self {
-        Auth {
-            username: username.to_owned(),
-            password: password.to_owned(),
-        }
-    }
-
-    pub fn to_bearer(&self) -> String {
-        format!("Bearer {}:{}", self.username, self.password)
-    }
-
-    pub fn username(&self) -> &str {
-        &self.username
-    }
-
-    pub fn password(&self) -> &str {
-        &self.password
-    }
-
-    pub fn parse(s: &str) -> Result<Self, ()> {
-        let s = s.replace("%3A", ":"); // When the token is received over the wire, it is URL encoded. We must decode the semi-colons.
-        if s.starts_with("Basic") {
-            Auth::parse_basic(&s[BASIC_TOKEN_START..])
-        } else if s.starts_with("Bearer") {
-            Auth::parse_bearer(&s[BEARER_TOKEN_START..])
-        } else {
-            Auth::parse_text(&s)
-        }
-    }
-
-    fn parse_basic(s: &str) -> Result<Self, ()> {
-        let decoded = base64::decode(&s).map_err(|_| ())?;
-        let text = std::str::from_utf8(&decoded).map_err(|_| ())?;
-        Auth::parse_text(text)
-    }
-
-    // Currently, we use bearer tokens in a non-standard way, where they each have a
-    // username and a password in them. In the future, we will deprecate Bearer auth
-    // for accounts.
-    fn parse_bearer(s: &str) -> Result<Auth, ()> {
-        Auth::parse_text(s)
-    }
-
-    fn parse_text(text: &str) -> Result<Self, ()> {
-        let parts = &mut text.split(':');
-        let username = match parts.next() {
-            Some(part) => part.to_owned(),
-            None => return Err(()),
-        };
-        let password = match parts.next() {
-            Some(part) => part.to_owned(),
-            None => return Err(()),
-        };
-        Ok(Auth { username, password })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parses_correctly() {
-        assert_eq!(
-            Auth::parse("Basic aW50ZXJsZWRnZXI6cnVzdA==").unwrap(),
-            Auth::new("interledger", "rust")
-        );
-
-        assert_eq!(
-            Auth::parse("Bearer interledger%3Arust").unwrap(),
-            Auth::new("interledger", "rust")
-        );
-
-        assert_eq!(
-            Auth::parse("Bearer interledger:rust").unwrap(),
-            Auth::new("interledger", "rust")
-        );
-
-        assert!(Auth::parse("SomethingElse asdf").is_err());
-        assert!(Auth::parse("Basic asdf").is_err());
-        assert!(Auth::parse("Bearer asdf").is_err());
-
-        assert_eq!(
-            Auth::new("interledger", "rust").to_bearer(),
-            "Bearer interledger:rust"
-        );
-    }
-
 }
