+++ crates/interledger-service/src/lib.rs
@@ -38,6 +38,9 @@ use std::{
 
 use serde::Serialize;
 
+mod auth;
+pub use auth::Auth as AuthToken;
+
 /// The base trait that Account types from other Services extend.
 /// This trait only assumes that the account has an ID that can be compared with others.
 ///
